pyproject.toml
[tool.poetry]
name = "django-belt"
version = "1.5.0"
description = "Simple package with some utilities for Django."
readme = "README.rst"
authors = ["Marcos Gabarda <hey@marcosgabarda.com>"]
license = "MIT"
homepage = "https://github.com/marcosgabarda/django-snitch"
packages = [
    { include = "belt" },
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "License :: OSI Approved :: MIT License",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.7"
django-model-utils = "^4.1.0"
djangorestframework = "^3.12.2"
django-filter = "^2.4.0"


[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
pytest = "^5.4.3"
black = "^20.8b1"
factory_boy = "^2.12.0"
django-test-plus = "^1.4.0"
pytest-django = "^3.9.0"
pytest-cov = "^2.10.0"
mypy = "^0.790"
isort = "^5.6.4"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
